+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2656 Fix AnalysisProfile keyword validator fail with non-ascii value 
+- #2646 Add SelectOtherField and SelectOtherWidget
 - #2649 Fix event subscribers are not triggered on analysis initialization
 - #2650 Fix AttributeError on analysis update with interim values via jsonapi
 - #2648 Increase the default width for field labels to min 150px